Logging changes to cached messages
Messages that existed before the bot was started will not have changes logged. In the future, a method to compare the new message's timestamp with $timestampSetup and skip messages with a timestamp older than $timestampSetup. This will be a more definite way to know in advance if nothing will be found in the message cache.