Leave invalid guilds
In the case that this bot is invited to a server where it does not have the ability to read messages, it will make a "best effort" attempt to leave the guild after criteria has been met. In this case, the bot keep track of updates to guild members. If at least 10 members in the guild update their accounts in a meaningful way (changing their avatar/username/discriminator) before it ever reads a mesage, the bot will leave the guild.