Remove periods from ckeys(?)
I'm not even sure this is needed, but there was a report of a user that had punctuation in their ckey. I need to verify whether or not that's the case, because if it is then I may need to urlencode the entire ckey instead of just removing the punctuation.